requirements.txt
alembic==1.10.4
blinker==1.6.2
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
Flask==2.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
googletrans==3.1.0a0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
pep8-naming==0.13.3
psycopg2==2.9.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
Werkzeug==2.3.1
WTForms==3.0.1
